+++ packages/terra-tabs/src/common-tabs/_Tab.jsx
@@ -130,6 +130,7 @@ const Tab = ({
     { 'is-active': isSelected },
     { 'is-icon-only': isIconOnly },
     { 'is-text-only': !icon },
+    { 'is-disabled': isDisabled },
     theme.className,
   );
   const paneClassNames = classNames(cy(
@@ -156,9 +157,12 @@ const Tab = ({
     }
   }
 
-  function onClick() {
+  function onClick(event) {
     if (!isDisabled) {
       onSelect(itemKey, metaData);
+      if (onChange) {
+        onChange(event, itemKey);
+      }
     }
   }
 
@@ -204,14 +208,15 @@ const Tab = ({
       id={id}
       aria-controls={associatedPanelId}
       role="tab"
+      aria-disabled={isDisabled}
       className={variant === 'framework' ? paneClassNames : tabClassNames}
       title={label}
       data-terra-tabs-show-focus-styles
     >
-      <div className={cx('inner')}>
+      <div className={variant === 'framework' ? cy('inner') : cx('inner')}>
         {customDisplay}
         {customDisplay ? null : icon}
-        {customDisplay || isIconOnly ? null : <span className={cx('label')}>{label}</span>}
+        {customDisplay || isIconOnly ? null : <span className={variant === 'framework' ? cy('label') : cx('label')}>{label}</span>}
       </div>
     </div>
   );